Additional modularization and adding a union type of all error codes
