+++ src/errorcode/all.errorcode.ts
@@ -1,0 +1,5 @@
+import { ProtocolErrorCode } from './protocol.errorcode';
+import { BioAuthErrorCode } from './bioauth.errorcode';
+import { SmsErrorCode } from './sms.errorcode';
+
+export type AllErrorCode = ProtocolErrorCode | BioAuthErrorCode | SmsErrorCode;
