+++ src/protocol.http.service.ts
@@ -1,7 +1,7 @@
 import { HttpService } from '@nestjs/common';
 import { AxiosRequestConfig, AxiosResponse } from 'axios';
 import { ProtocolException } from './protocol.exception';
-import { ProtocolErrorCode } from './protocol.errorcode';
+import { ProtocolErrorCode } from './errorcode/protocol.errorcode';
 import { Observable } from 'rxjs';
 import { RequestContext } from './http-context/request.context';
 import { Logger } from './logger';
